Rename home to DistanceConversion
